Fix protobufDefinitionsToDatatypes for Timestamp/Duration (#6785)
**User-Facing Changes**
Fixed an incompatibility between timestamp/duration types in Protobuf
messages and User Scripts.

**Description**
Fixes #6783 
Fixes FG-4764

Background:

Protobuf's standard Timestamp and Duration types use int64 `seconds` +
int32 `nanos`, but most of Studio works with timestamps in `{sec,nsec}`
format with `number` values.

Prior work:

- In #3443 I added code to modify these definitions to use `sec`/`nsec`
instead for compatibility with the rest of Studio — however, I neglected
to change the types in User Scripts.

- In #5727 I added `bigint` support to our Protobuf deserialization,
which accidentally changed these timestamps to use `bigint` for `sec`,
which I fixed in #5757.

- In #6249 I changed the User Script generated types to use the
`sec`/`nsec` field names, but I didn't realize that they still used a
`bigint` type for the `sec` field.

This PR finally fixes the User Script types so that they have `sec:
number` instead of `sec: bigint`.

+++ packages/mcap-support/src/parseProtobufSchema.test.ts
@@ -200,7 +200,7 @@ describe("parseProtobufSchema", () => {
           {
             "isArray": false,
             "name": "sec",
-            "type": "int64",
+            "type": "int32",
           },
           {
             "isArray": false,
@@ -216,7 +216,7 @@ describe("parseProtobufSchema", () => {
           {
             "isArray": false,
             "name": "sec",
-            "type": "int64",
+            "type": "int32",
           },
           {
             "isArray": false,
